FIX STRTAB INCORRECT OFFSET IN SYMBOLS:
In case the symbol points on an incorrect section, now check if
the section is of type SHT_STRTAB.
Thus printing (null) as name in case it's not a strtab.

There is still an error in type recognition in this case, but I will
not be fixing this as it's in my opinion not worth spending more time on
it.